+++ database/migrations/2025_02_11_111643_create_projects_table.php
@@ -17,9 +17,9 @@ return new class extends Migration
             $table->string('title');
             $table->text('description')->nullable();
             $table->string('project_icon')->nullable();
-            $table->integer('taxonomic_scope_id');
-            $table->integer('created_by_id')->nullable();
-            $table->integer('updated_by_id')->nullable();
+            $table->unsignedBigInteger('taxonomic_scope_id');
+            $table->unsignedBigInteger('created_by_id')->nullable();
+            $table->unsignedBigInteger('updated_by_id')->nullable();
             $table->index('title');
             $table->index('taxonomic_scope_id');
             $table->index('created_by_id');
