+++ database/migrations/2025_02_11_114521_create_items_table.php
@@ -17,9 +17,9 @@ return new class extends Migration
             $table->uuid('guid');
             $table->string('name');
             $table->string('url')->nullable();
-            $table->integer('project_id');
-            $table->integer('created_by_id')->nullable();
-            $table->integer('updated_by_id')->nullable();
+            $table->unsignedBigInteger('project_id');
+            $table->unsignedBigInteger('created_by_id')->nullable();
+            $table->unsignedBigInteger('updated_by_id')->nullable();
             $table->index('name');
             $table->index('project_id');
             $table->index('created_by_id');
